feat: implement applyLimit function to restrict repeater items
 update BlockRenderer to utilize it
